update /me url

+++ src/main/java/enzocesarano/GestioneEventi/controllers/UtenteController.java
@@ -4,6 +4,8 @@ import enzocesarano.GestioneEventi.entities.Evento;
 import enzocesarano.GestioneEventi.entities.Prenotazione;
 import enzocesarano.GestioneEventi.entities.Utente;
 import enzocesarano.GestioneEventi.exceptions.BadRequestException;
+import enzocesarano.GestioneEventi.exceptions.NotFoundException;
+import enzocesarano.GestioneEventi.exceptions.UnauthorizedException;
 import enzocesarano.GestioneEventi.payloads.EventoDTO;
 import enzocesarano.GestioneEventi.payloads.PrenotazioneDTO;
 import enzocesarano.GestioneEventi.payloads.UtenteDTO;
@@ -51,67 +53,90 @@ public class UtenteController {
     }
 
 
-    @GetMapping("/{id_utente}/prenotazioni")
+    @GetMapping("/me/prenotazioni")
     @ResponseStatus(HttpStatus.OK)
-    public List<Prenotazione> getPrenotazioniByUtente(@PathVariable UUID id_utente) {
-        List<Prenotazione> prenotazioni = this.utenteService.findPrenotazioniByUtenteId(id_utente);
+    public List<Prenotazione> getPrenotazioniByUtente(@AuthenticationPrincipal Utente currentAuthenticatedUtente) {
+        List<Prenotazione> prenotazioni = this.utenteService.findPrenotazioniByUtenteId(currentAuthenticatedUtente.getId_utente());
         return prenotazioni;
     }
 
-    @PostMapping("/{id_utente}/prenotazioni")
+    @PostMapping("/me/prenotazioni")
     @ResponseStatus(HttpStatus.CREATED)
-    public Prenotazione postPrenotazione(@RequestBody @Validated PrenotazioneDTO payload, @PathVariable UUID id_utente, BindingResult validationResult) {
+    public Prenotazione postPrenotazione(@RequestBody @Validated PrenotazioneDTO payload, @AuthenticationPrincipal Utente currentAuthenticatedUtente, BindingResult validationResult) {
         if (validationResult.hasErrors()) {
             validationResult.getAllErrors().forEach(System.out::println);
             throw new BadRequestException("Ci sono stati errori nel payload!");
         }
-        return this.prenotazioneService.savePrenotazione(payload, id_utente);
+        return this.prenotazioneService.savePrenotazione(payload, currentAuthenticatedUtente.getId_utente());
     }
 
-    @GetMapping("/{id_utente}/eventi")
+    @GetMapping("/me/eventi")
     @ResponseStatus(HttpStatus.OK)
     @PreAuthorize("hasAuthority('ORGANIZZATORE_EVENTI')")
     public Page<Evento> findAllEventiPerUtente(
-            @PathVariable UUID id_utente,
+            @AuthenticationPrincipal Utente currentAuthenticatedUtente,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size,
             @RequestParam(defaultValue = "dataEvento") String sortBy) {
-        return this.eventoService.findAllByOrganizzatore(id_utente, page, size, sortBy);
+        return this.eventoService.findAllByOrganizzatore(currentAuthenticatedUtente.getId_utente(), page, size, sortBy);
     }
 
-    @GetMapping("/{id_utente}/eventi/{id_evento}")
+    @GetMapping("/me/eventi/{id_evento}")
     @ResponseStatus(HttpStatus.OK)
-    @PreAuthorize("hasAuthority('ORGANIZZATORE_EVENTI')")
-    public Evento getEvento(@PathVariable UUID id_evento) {
-        return this.eventoService.findById(id_evento);
+    @PreAuthorize("hasAuthority('ROLE_ORGANIZZATORE_EVENTI')")
+    public Evento getEvento(@AuthenticationPrincipal Utente currentAuthenticatedUtente, @PathVariable UUID id_evento) {
+        Evento evento = this.eventoService.findById(id_evento);
+        if (evento == null || !evento.getOrganizzatore().equals(currentAuthenticatedUtente)) {
+            throw new NotFoundException("Evento non trovato o non appartiene all'utente!");
+        }
+        return evento;
     }
 
-    @PostMapping("/{id_utente}/eventi")
+    @PostMapping("/me/eventi")
     @ResponseStatus(HttpStatus.CREATED)
     @PreAuthorize("hasAuthority('ORGANIZZATORE_EVENTI')")
-    public Evento postEvento(@RequestBody @Validated EventoDTO payload, @PathVariable UUID id_utente, BindingResult validationResult) {
+    public Evento postEvento(@RequestBody @Validated EventoDTO payload, @AuthenticationPrincipal Utente currentAuthenticatedUtente, BindingResult validationResult) {
         if (validationResult.hasErrors()) {
             validationResult.getAllErrors().forEach(System.out::println);
             throw new BadRequestException("Ci sono stati errori nel payload!");
         }
-        return this.eventoService.saveEvento(payload, id_utente);
+        return this.eventoService.saveEvento(payload, currentAuthenticatedUtente.getId_utente());
     }
 
-    @PutMapping("/{id_utente}/eventi/{id_evento}")
+    @PutMapping("/me/eventi/{id_evento}")
     @ResponseStatus(HttpStatus.OK)
-    @PreAuthorize("hasAuthority('ORGANIZZATORE_EVENTI')")
-    public Evento putEvento(@PathVariable UUID id_evento, @RequestBody @Validated EventoDTO payload, BindingResult validationResult) {
+    @PreAuthorize("hasAuthority('ROLE_ORGANIZZATORE_EVENTI')")
+    public Evento putEvento(@AuthenticationPrincipal Utente currentAuthenticatedUtente,
+                            @PathVariable UUID id_evento,
+                            @RequestBody @Validated EventoDTO payload,
+                            BindingResult validationResult) {
         if (validationResult.hasErrors()) {
             validationResult.getAllErrors().forEach(System.out::println);
             throw new BadRequestException("Ci sono stati errori nel payload!");
+        }
+        Evento evento = this.eventoService.findById(id_evento);
+        if (evento == null) {
+            throw new NotFoundException("Evento non trovato!");
+        }
+
+        if (!evento.getOrganizzatore().equals(currentAuthenticatedUtente)) {
+            throw new UnauthorizedException("Non hai i permessi per aggiornare questo evento!");
         }
         return this.eventoService.findByIdAndUpdate(id_evento, payload);
     }
 
-    @DeleteMapping("/{id_utente}/eventi/{id_evento}")
+    @DeleteMapping("/me/eventi/{id_evento}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PreAuthorize("hasAuthority('ORGANIZZATORE_EVENTI')")
-    public void deleteEvento(@PathVariable UUID id_evento) {
+    @PreAuthorize("hasAuthority('ROLE_ORGANIZZATORE_EVENTI')")
+    public void deleteEvento(@AuthenticationPrincipal Utente currentAuthenticatedUtente,
+                             @PathVariable UUID id_evento) {
+        Evento evento = this.eventoService.findById(id_evento);
+        if (evento == null) {
+            throw new NotFoundException("Evento non trovato!");
+        }
+        if (!evento.getOrganizzatore().equals(currentAuthenticatedUtente)) {
+            throw new UnauthorizedException("Non hai i permessi per eliminare questo evento!");
+        }
         this.eventoService.deleteEvento(id_evento);
     }
 }
